Rename concurrency module to 'primitive'
